Moved logic functions out of UI component
